Revert "h_signature: Use double rather than long double arithmetic"
This reverts commit 6b9b00d41e1bdd9da80c234fac7adf9f29bd3951.

+++ include/teb_local_planner/equivalence_relations.h
@@ -63,7 +63,7 @@ namespace teb_local_planner
 class EquivalenceClass
 {
 public:
-   typedef std::complex<double> cplx;
+   typedef std::complex<long double> cplx;
 
    /**
     * @brief Default constructor
